requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dominate==2.9.1
Flask==3.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
numpy==1.26.2
openai==1.12.0
packaging==24.0
pillow==10.3.0
plotly==5.18.0
pydantic==2.7.2
pydantic_core==2.18.3
PyMySQL==1.1.1
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.3
six==1.16.0
sniffio==1.3.1
spotipy==2.23.0
SQLAlchemy==2.0.30
tenacity==8.3.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
visitor==0.1.3
Werkzeug==3.0.3
WTForms==3.1.2
